fix syntax-error on older python interpreters
