+++ tox.ini
@@ -17,6 +17,14 @@ commands =
     {posargs}
   {env:CI_COMMANDS:python -c ""}
 
+[testenv:py35]
+whitelist_externals =
+  bash
+deps =
+  -rrequirements.txt
+commands =
+  bash -c "mitmdump --version 2>&1 | grep 'mitmproxy requires Python 3.6'"
+
 [testenv:lint]
 commands =
   mitmdump --version
